Implement add (#1)
* First draft of add. Still some bugs

* Actually hash the object and fix name length bug

+++ src/commands/remove.rs
@@ -1,9 +1,12 @@
-use std::{env, fs};
+use std::fs;
 
 use anyhow::Context;
 use clap::Args;
 
-use crate::{objects::Index, util::repo_root};
+use crate::{
+    objects::Index,
+    util::normalize_paths,
+};
 
 #[derive(Args, Debug)]
 pub struct RemoveArg {
@@ -12,39 +15,33 @@ pub struct RemoveArg {
 
 impl RemoveArg {
     pub fn run(self) -> anyhow::Result<()> {
-        let root = repo_root()?;
+        let new_index = remove(&self.paths, true)?;
 
-        let index = Index::parse_from_disk()?;
-
-        let cwd = env::current_dir()?;
-        let normalized_paths = self
-            .paths
-            .iter()
-            .map(|path| -> anyhow::Result<String> {
-                let absolute_path = cwd.join(path).canonicalize()?;
-                let repo_relative_path = absolute_path.strip_prefix(&root).with_context(|| {
-                    format!("Path {absolute_path:?} is not within the repository")
-                })?;
-
-                Ok(repo_relative_path.to_string_lossy().into())
-            })
-            .collect::<anyhow::Result<Vec<String>>>()?;
-
-        // TODO: This is inefficient, we should be able to do this in one pass
-        let new_entries = index
-            .entries
-            .into_iter()
-            .filter(|entry| !normalized_paths.contains(&entry.name))
-            .collect::<Vec<_>>();
-
-        let bytes = Index::from_entries(new_entries).serialize()?;
-
-        for p in self.paths.iter() {
-            fs::remove_file(p).with_context(|| format!("Failed to remove file '{}'", p))?;
-        }
-
-        fs::write(root.join(".bit/index"), bytes).context("Failed to write new index to disk")?;
+        new_index.write_to_disk()?;
 
         Ok(())
     }
+}
+
+pub fn remove(paths: &[String], delete_file: bool) -> anyhow::Result<Index> {
+    let index = Index::parse_from_disk()?;
+
+    let normalized_paths = normalize_paths(paths.as_ref())?;
+
+    // TODO: This is inefficient, we should be able to do this in one pass
+    let new_entries = index
+        .entries
+        .into_iter()
+        .filter(|entry| !normalized_paths.contains(&entry.name))
+        .collect::<Vec<_>>();
+
+    let new_index = Index::from_entries(new_entries);
+
+    if delete_file {
+        for p in paths.iter() {
+            fs::remove_file(p).with_context(|| format!("Failed to remove file '{}'", p))?;
+        }
+    }
+
+    Ok(new_index)
 }
